chore(web): style cleanup related to outline, show edge in deletion confirmation modal
